requirements.txt
discord.py>=2.4.0
python-dotenv
aiohttp
tenacity
croniter
tavily-python
beautifulsoup4
requests
colorama
